+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swap/tradeoptions/RecipientAddressViewModel.kt
@@ -98,7 +98,7 @@ class RecipientAddressViewModel(
         setTextLiveData.postValue(addressData.address)
 
         addressData.amount?.let {
-            service.set(it.toBigDecimal())
+            service.set(it)
         }
     }
